+++ prompt_styles/light-double-line.sh
@@ -63,7 +63,7 @@ fancygit_prompt_builder() {
 
     prompt_user="${user_at_host}\\u@\\h ${user_at_host_end}"
     prompt_symbol="\n${user_symbol}\$${user_symbol_end}"
-    prompt_path="${path}${bold}${white} \\w ${path_end}${s_darkgray}"
+    prompt_path="${path}${bold}${black} \\w ${path_end}${s_lightgray}"
 
     local only_local_branch=$(git branch -a 2> /dev/null | egrep "remotes/origin/${branch_name}" | wc -l)
 
@@ -79,7 +79,16 @@ fancygit_prompt_builder() {
         return
     fi
 
-    PS1="${prompt_user}${prompt_path}${prompt_symbol} "
+    local venv=""
+
+    if ! [ -z ${VIRTUAL_ENV} ]
+    then
+        prompt_user="${user_at_host} \\u@\\h ${user_at_host_end}"
+        venv="`basename \"$VIRTUAL_ENV\"`"
+        venv="${bg_light_gray}${bold}${black}${venv} ${path_end}${s_lightgray_bgwhite}"
+    fi
+
+    PS1="${venv}${prompt_user}${prompt_path}${prompt_symbol} "
 }
 
 PROMPT_COMMAND="fancygit_prompt_builder"